fixed error handling
